Moved all scripts to one folder. New fil Fig4_S4-5-scripts.tar.gz replaces Fig3_S5_9-scripts.tar.gz
