+++ results/scripts/annotation/maker_exe.ctl
@@ -1,0 +1,26 @@
+#-----Location of Executables Used by MAKER/EVALUATOR
+makeblastdb=/ebio/abt6/alkeller/workspace/software/blast/bin/makeblastdb #location of NCBI+ makeblastdb executable
+blastn=/ebio/abt6/alkeller/workspace/software/blast/bin/blastn #location of NCBI+ blastn executable
+blastx=/ebio/abt6/alkeller/workspace/software/blast/bin/blastx #location of NCBI+ blastx executable
+tblastx=/ebio/abt6/alkeller/workspace/software/blast/bin/tblastx #location of NCBI+ tblastx executable
+formatdb=/usr/bin/formatdb #location of NCBI formatdb executable
+blastall=/usr/bin/blastall #location of NCBI blastall executable
+xdformat= #location of WUBLAST xdformat executable
+blasta= #location of WUBLAST blasta executable
+prerapsearch= #location of prerapsearch executable
+rapsearch= #location of rapsearch executable
+RepeatMasker=/ebio/abt6_projects9/abt6_software/bin/repeatmasker/RepeatMasker #location of RepeatMasker executable
+exonerate=/ebio/abt6_projects9/abt6_software/bin/exonerate/bin/exonerate #location of exonerate executable
+
+#-----Ab-initio Gene Prediction Algorithms
+snap= /ebio/abt6_projects9/abt6_software/bin/snap/snap #location of snap executable
+gmhmme3= #location of eukaryotic genemark executable
+gmhmmp= #location of prokaryotic genemark executable
+augustus=/ebio/abt6_projects9/abt6_software/bin/augustus/bin/augustus #location of augustus executable
+fgenesh= #location of fgenesh executable
+evm=#/ebio/abt6_projects7/small_projects/alkeller/software/evm/evidence_modeler.pl  #location of EvidenceModeler executable
+tRNAscan-SE= #location of trnascan executable
+snoscan= #location of snoscan executable
+
+#-----Other Algorithms
+probuild= #location of probuild executable (required for genemark)
